Hi Bamvor,
Attached is my latest version of a patch to fix the libxl races.  It is
compile-tested only!!  I'm too tired to start testing it now, but
thought I would pass it along to you for review/comment.  I'll start
testing it first thing tomorrow morning.

Unfortunately, we need to keep a list of all registered timeouts and
fds, and ensure they are deregistered (removed from libvirt's event
       loop) before freeing the associated libxlDomainObjPrivate object.  If
not, when the libxl_ctx is freed, fd's and timeouts that are still
registered can have their callbacks invoked, which is not good since
everything is being freed, including the libxl_ctx :).

This patch keeps a list of registered timeouts and fds in the
libxlDomainObjPrivate struct.  The info struct associated with timeouts
and fds is also reference counted to ensure it is not freed by the
activities of another thread.

Even if the patch worked as is, it would needs some adjustments before
sending upstream.  E.g. the code to add/remove items from the list can
probably be macro-ized.  But that shouldn't prevent a review of the
general approach to fixing the races, so any thoughts you might have
would be appreciated.

Jim